Fixed #14
- Added BlockingQueue to support proper cancellation after queue has
been disposed of
- Ensured that pending messages from a previous connection are NOT
mistakenly sent over the new connection
- Ensured that disconnects due to errors with the previous connection
are NOT mistakenly used to disconnect the next connection

+++ SharpRemote.Windows/ConnectionId.cs
@@ -44,9 +44,9 @@ namespace SharpRemote
 			return !left.Equals(right);
 		}
 
-		private readonly long _value;
+		private readonly int _value;
 
-		internal ConnectionId(long value)
+		internal ConnectionId(int value)
 		{
 			_value = value;
 		}
